requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
praw==7.1.0
prawcore==1.4.0
python-decouple==3.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
update-checker==0.17
urllib3==1.25.10
vk-api==11.8.0
websocket-client==0.57.0
